test(unit): update specs for SendCode components
